genefinding: move from 'detection' namespace to 'support'
