+++ antismash/support/genefinding/data/train_crypto/acc1.mar
