+++ antismash/support/genefinding/data/train_crypto/atg.dt
