+++ antismash/support/genefinding/data/train_crypto/coding_0_100.model
