+++ antismash/support/genefinding/data/train_crypto/config.file
