+++ antismash/support/genefinding/data/train_crypto/exon.distr
