+++ antismash/support/genefinding/data/train_crypto/exons.dat
