+++ antismash/support/genefinding/data/train_crypto/false.acc
