+++ antismash/support/genefinding/data/train_crypto/noncoding_0_100.model
