+++ antismash/support/genefinding/data/train_crypto/orfs_0_100
