+++ antismash/support/genefinding/data/train_crypto/outex
