+++ antismash/support/genefinding/data/train_crypto/score_ex.acc
