+++ antismash/support/genefinding/data/train_crypto/seqs.out
