+++ antismash/support/genefinding/data/train_crypto/seqs_0_100
