+++ antismash/support/genefinding/data/train_crypto/sites.stop
