+++ antismash/support/genefinding/data/train_crypto/stop.markov
