+++ antismash/support/genefinding/data/train_crypto/temp.fasta
