+++ antismash/support/genefinding/data/train_crypto/train_0_100.cfg
