+++ antismash/support/genefinding/run_prodigal.py
