+++ antismash/support/genefinding/test/data/fumigatus.cluster1.fna
