+++ antismash/support/genefinding/test/integration_prodigal.py
@@ -9,7 +9,7 @@ from unittest import TestCase
 from antismash.config import get_config, update_config
 from antismash.common.record_processing import parse_input_sequence, pre_process_sequences
 from antismash.common.test.helpers import get_simple_options, get_path_to_nisin_fasta
-from antismash.detection import genefinding
+from antismash.support import genefinding
 
 
 class TestProdigal(TestCase):
